+++ src/page/Editor/hooks/useResizeWidget.ts
@@ -1,6 +1,6 @@
 import { useCallback } from "react"
 import { useDispatch } from "react-redux"
-import { widgetStatesActions } from "@/redux/reducers/editorReducer/widgetStatesReducer"
+import { widgetStatesActions } from "@/redux/editor/widgetStatesReducer"
 
 export const useResizeWidget = () => {
   const dispatch = useDispatch()
