+++ src/redux/action/actionList/actionListSelector.ts
@@ -1,0 +1,6 @@
+import { EditorState } from "@/redux/action/editor/editorReducer"
+
+export interface ActionState {
+  editor: EditorState
+  queryList: QueryListState
+}
