+++ src/redux/action/editor/editorReducer.ts
