+++ src/redux/dashboard/dashboardSelector.ts
@@ -1,0 +1,3 @@
+import { createSelector } from "@reduxjs/toolkit"
+
+export const getDashboardApps = createSelector(() => {})
