+++ src/redux/editor/dragReducer.ts
