+++ src/redux/editor/editorSeletor.ts
