+++ src/redux/editor/widgetStatesReducer.ts
