--- src/redux/reducers/interface.ts
@@ -1,9 +1,0 @@
-import { EditorReduxState } from "./editorReducer"
-import { ActionState } from "./actionReducer"
-
-export interface BuilderState {
-  editor: EditorReduxState
-  action: ActionState
-  userGroup: {}
-  metaInfo: {}
-}
