--- src/redux/reducers/metaInfoReducer/index.ts
@@ -1,5 +1,0 @@
-import { combineReducers } from "redux"
-
-const metaInfoReducer = ""
-
-export default metaInfoReducer
