--- src/redux/reducers/userGroupReducer/index.ts
@@ -1,5 +1,0 @@
-import { combineReducers } from "redux"
-
-const userGroupReducer = ""
-
-export default userGroupReducer
