--- src/redux/sagas/demo2.ts
@@ -1,15 +1,0 @@
-import { all, put, takeLatest, delay } from "redux-saga/effects"
-import { changeDemoValueA } from "../reducers/editorReducer/demoReducer"
-
-function* incrementAsync() {
-  yield delay(1000)
-  yield put(changeDemoValueA())
-}
-function* watchIncrement() {
-  yield takeLatest("incrementAsync_demo_saga", incrementAsync)
-}
-
-function* demo2Saga() {
-  yield all([watchIncrement()])
-}
-export default demo2Saga
